Add files via upload
This is  a library application which allows the users to save, edit and delete Paper books, e-books, and audiobooks. In part 2, we will design the
reading, deleting, and editing functionalities. When user navigates to “Read and Edit”, all books are being shown in the next
form (You need to get this data from the static list in your static class). user can navigate between the books and see them. The label at the middle shows
the current number of the book in the list and the total number of available books in the list.